Refactor: LoginVO.java 폴더 이동
java user와 맞추어 구조 변경
