--- src/main/java/kr/co/solfood/login/LoginVO.java
@@ -1,19 +1,0 @@
-package kr.co.solfood.login;
-
-import lombok.Data;
-
-@Data
-public class LoginVO {
-    private long usersId;         // PK
-    private int companyId;       // FK
-    private int departmentId;    // FK
-    private String usersEmail;      // Unique or FK
-    private String usersNickname;   // Unique or FK
-    private String usersProfile;
-    private String usersName;
-    private int usersPoint;
-    private int usersAge;
-    private String usersGender;
-    private String accessToken;
-    private long usersKakaoId;
-}
