+++ src/main/java/kr/co/solfood/user/login/LoginVO.java
@@ -1,0 +1,23 @@
+package kr.co.solfood.user.login;
+
+import lombok.Data;
+
+@Data
+public class LoginVO {
+    private long usersId;
+    private int companyId;
+    private int departmentId;
+    private String usersEmail;
+    private String usersNickname;
+    private String usersProfile;
+    private String usersName;
+    private int usersPoint;
+    private String usersGender;
+    private String accessToken;
+    private long usersKakaoId;
+    private String usersTel;
+    private String usersJoinDate;
+    private String usersEnterDate;
+    private String usersStatus;
+    private String usersLoginType;
+}
